Web release! (#56)
- Fixed crash when editing a message
- Fixed broken price queries
- Finalized logos, title, and colors
- removed hard-coding from user profile page
- edit profile page
- made seller page responsive
- added disclaimer to product editor
- added a dark mode 
- added settings for dark mode to user profile
- removed all subclasses of `ProductFilters`
- persist user settings like archived conversations when saving account
- all conversations are now managed through the `ConversationsModel`
- all data models now implement `onSignIn(UserProfile)` and
`onSignOut()`
- In-app notifications
- reduced queries by limiting responses
- added an error page
- fixed a bug with buyer images

---------

+++ lib/src/models/builders/seller_profile.dart
@@ -47,7 +47,6 @@ class SellerProfileBuilder extends BuilderModel<SellerProfile> {
 
   /// The seller's ID.
   late final SellerID sellerID;
-  // SellerID get sellerID => SellerID("123");
 
   /// Fetching the email address of the seller
   String get email => services.auth.user!.email!;
@@ -90,7 +89,7 @@ class SellerProfileBuilder extends BuilderModel<SellerProfile> {
       twitterController.text = seller.contact.twitterUsername != null ? seller.contact.twitterUsername! : "";
       linkedinController.text = seller.contact.linkedInUsername != null ? seller.contact.linkedInUsername! : "";
     }
-    userID = models.user.userProfile!.id;
+    userID = models.user.userID!;
     for (final controller in allControllers) {
       controller.addListener(notifyListeners);
     }
@@ -108,10 +107,10 @@ class SellerProfileBuilder extends BuilderModel<SellerProfile> {
   @override
   SellerProfile build() => SellerProfile(
     id: sellerID,
-    name: nameController.text,
+    name: nameController.text.trim(),
     userID: userID,
     imageUrl: imageUrl!,
-    bio: bioController.text,
+    bio: bioController.text.trim(),
     contact: ContactInfo(
       email: email,
       phoneNumber: phoneNumberController.text.nullIfEmpty,
@@ -148,7 +147,7 @@ class SellerProfileBuilder extends BuilderModel<SellerProfile> {
 
   /// Deletes the image at the given index.
   Future<void> deleteImage() async {
-    // [filename] cannot be null because imageUrl is required for seller profile, and we're copying it before nullifying it.
+    if (imageUrl == null) return;
     await services.cloudStorage.deleteFile(imageUrl!);
     imageUrl = null;
     notifyListeners();
@@ -166,8 +165,7 @@ class SellerProfileBuilder extends BuilderModel<SellerProfile> {
     saveError = null;
     try {
       final result = build();
-      await services.database.saveSellerProfile(result);
-      await models.user.loadSellerProfiles();
+      await models.user.addSellerProfile(result);
       if (isEditing) {
         router.pop(result);
       } else {
